refactor: modify fields

+++ src/service/serializers/event_serializers/CreateEventSerializer.py
@@ -9,7 +9,7 @@ class CreateEventSerializer(serializers.Serializer):
     attendance = serializers.CharField(max_length = 1)
     maximum_tickets = serializers.IntegerField()
     is_paid = serializers.BooleanField()
-    ticket_price = serializers.DecimalField(max_digits = 6, decimal_places = 2, required = False)
+    ticket_price = serializers.IntegerField(required = False)
     organizer_phone = serializers.CharField(max_length = 11)
     organizer_SSN = serializers.CharField(max_length = 11)
     photo = serializers.ImageField()
@@ -18,8 +18,8 @@ class CreateEventSerializer(serializers.Serializer):
     province = serializers.CharField(max_length = 255, required = False)
     city = serializers.CharField(max_length = 255, required = False)
     address = serializers.CharField(max_length = 255, required = False)
-    location_lat = serializers.DecimalField(max_digits = 9, decimal_places = 6, required = False)
-    location_lon = serializers.DecimalField(max_digits = 9, decimal_places = 6, required = False)
+    location_lat = serializers.DecimalField(max_digits = 17, decimal_places = 15, required = False)
+    location_lon = serializers.DecimalField(max_digits = 18, decimal_places = 15, required = False)
 
     tags = serializers.ListField(required = False)
