+++ src/service/serializers/event_serializers/EventSummarySerializer.py
@@ -8,7 +8,7 @@ class EventSummarySerializer(serializers.ModelSerializer):
     category = serializers.CharField(max_length = 255)
     attendance = serializers.CharField(max_length = 1)
     is_paid = serializers.BooleanField()
-    ticket_price = serializers.DecimalField(max_digits = 6, decimal_places = 2, required = False)
+    ticket_price = serializers.IntegerField(required = False)
     start_date = serializers.DateField()
     photo = serializers.ImageField()
     province = serializers.CharField(max_length = 255, source = "inpersonevent.province")
